# Migrate `torch.device` bindings to pybind11