--- a/torch/csrc/Device.cpp
+++ b/torch/csrc/Device.cpp
@@ -1,235 +1,106 @@
 #include <torch/csrc/Device.h>
-
 #include <torch/csrc/Exceptions.h>
-#include <torch/csrc/utils/object_ptr.h>
-#include <torch/csrc/utils/python_arg_parser.h>
-#include <torch/csrc/utils/python_strings.h>
-#include <torch/csrc/utils/python_numbers.h>
-#include <torch/csrc/utils/pybind.h>
 
-#include <ATen/Device.h>
-#include <c10/util/Exception.h>
-
-#include <cstring>
-#include <limits>
-#include <structmember.h>
 #include <sstream>
 
-PyObject *THPDevice_New(const at::Device& device)
-{
-  auto type = (PyTypeObject*)&THPDeviceType;
-  auto self = THPObjectPtr{type->tp_alloc(type, 0)};
-  if (!self) throw python_error();
-  auto self_ = reinterpret_cast<THPDevice*>(self.get());
-  self_->device = device;
-  return self.release();
+namespace torch {
+using namespace at;
+
+namespace {
+
+inline Device parseDeviceIndex(DeviceType type, DeviceIndex index) {
+  // -1 is allowed in ATen/C++, to mean the default device, but not in Python
+  TORCH_CHECK(index >= 0, "Device index must not be negative");
+  return Device(type, index);
 }
 
-PyObject *THPDevice_repr(THPDevice *self)
-{
+inline std::string deviceType(const Device& device) {
   std::ostringstream oss;
-  oss << "device(type=\'" << self->device.type() << "\'";
-  if (self->device.has_index()) {
-    // `self->device.index()` returns uint8_t which is treated as ascii while printing,
-    // hence casting it to uint16_t.
-    // https://stackoverflow.com/questions/19562103/uint8-t-cant-be-printed-with-cout
-    oss << ", index=" << static_cast<uint16_t>(self->device.index());
-  }
-  oss << ")";
-  return THPUtils_packString(oss.str().c_str());
+  oss << device.type();
+  return oss.str();
 }
 
-PyObject *THPDevice_str(THPDevice *self)
-{
-  std::ostringstream oss;
-  oss << self->device;
-  return THPUtils_packString(oss.str().c_str());
-}
+} // namespace
 
-PyObject *THPDevice_pynew(PyTypeObject *type, PyObject *args, PyObject *kwargs)
-{
-  HANDLE_TH_ERRORS
-  static torch::PythonArgParser parser({
-    "Device(Device device)",
-    "Device(std::string type, int64_t? index=-1)"
-  });
-  torch::ParsedArgs<2> parsed_args;
-  auto r = parser.parse(args, kwargs, parsed_args);
-  if (r.idx == 0) {
-    auto device = r.device(0);
-    return THPDevice_New(device);
-  } else if (r.idx == 1) {
-    auto as_device = r.device(0);  // this works, because device can take strings
-    auto device_type = r.string(0);
-    if (as_device.has_index()) {
-      throw std::runtime_error("type (string) must not include an index because index "
-                                "was passed explicitly: " + device_type);
-    }
-    int32_t device_index = -1;
-    if (!r.isNone(1)) {
-      device_index = r.toInt64(1);
-      // -1 is allowed in ATen/C++, to mean the default device, but not in
-      // Python.
-      TORCH_CHECK(device_index >= 0, "Device index must not be negative");
-    }
-    at::Device device(as_device.type(), device_index);
-    return THPDevice_New(device);
-  }
-  Py_RETURN_NONE;
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject *THPDevice_type(THPDevice *self, PyObject *noargs)
-{
-  HANDLE_TH_ERRORS
-  std::ostringstream oss;
-  oss << self->device.type();
-  return THPUtils_packString(oss.str().c_str());
-  Py_RETURN_NONE;
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject *THPDevice_index(THPDevice *self, PyObject *noargs)
-{
-  HANDLE_TH_ERRORS
-  if (self->device.has_index()) {
-    return THPUtils_packInt64(self->device.index());
+Device parseDevice(py::object device) {
+  if (py::isinstance<Device>(device)) {
+    return device.cast<Device>();
+  } else if (py::isinstance<py::str>(device)) {
+    return Device(device.cast<std::string>());
+  } else if (py::isinstance<py::int_>(device)) {
+    return parseDeviceIndex(kCUDA, device.cast<DeviceIndex>());
   } else {
-    Py_RETURN_NONE;
-  }
-  END_HANDLE_TH_ERRORS
-}
-
-static Py_ssize_t THPDevice_hash(THPDevice *self)
-{
-  HANDLE_TH_ERRORS
-  return static_cast<Py_ssize_t>(std::hash<at::Device>{}(self->device) % std::numeric_limits<Py_ssize_t>::max());
-  END_HANDLE_TH_ERRORS_RET(-1)
-}
-
-PyObject *THPDevice_rc(PyObject *a, PyObject *b, int op) {
-  HANDLE_TH_ERRORS
-  if (!THPDevice_Check(a) || !THPDevice_Check(b)) {
-    // Py_RETURN_NOTIMPLEMENTED not in python 2.
-    Py_INCREF(Py_NotImplemented);
-    return Py_NotImplemented;
-  }
-  THPDevice *da = reinterpret_cast<THPDevice*>(a);
-  THPDevice *db = reinterpret_cast<THPDevice*>(b);
-
-  switch(op) {
-    case Py_EQ:
-      if (da->device == db->device) {
-        Py_RETURN_TRUE;
-      } else {
-        Py_RETURN_FALSE;
-      }
-    case Py_NE:
-      if (da->device == db->device) {
-        Py_RETURN_FALSE;
-      } else {
-        Py_RETURN_TRUE;
-      }
-    case Py_LT:
-    case Py_LE:
-    case Py_GT:
-    case Py_GE:
-      throw torch::TypeError("comparison not implemented");
-    default:
-      throw torch::TypeError("unexpected comparison op");
-  }
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject *THPDevice_reduce(PyObject *_self, PyObject *noargs)
-{
-  HANDLE_TH_ERRORS
-  auto self = (THPDevice*)_self;
-  auto ret = THPObjectPtr{PyTuple_New(2)};
-  if (!ret) throw python_error();
-
-  py::object torch_module = py::module::import("torch");
-  py::object torch_device = torch_module.attr("device");
-  PyTuple_SET_ITEM(ret.get(), 0, torch_device.release().ptr());
-
-  THPObjectPtr args;
-  std::ostringstream oss;
-  oss << self->device.type();
-  if (self->device.has_index()) {
-    args = THPObjectPtr{Py_BuildValue("(si)", oss.str().c_str(), self->device.index())};
-  } else {
-    args = THPObjectPtr{Py_BuildValue("(s)", oss.str().c_str())};
-  }
-  if (!args) throw python_error();
-  PyTuple_SET_ITEM(ret.get(), 1, args.release());
-
-  return ret.release();
-  END_HANDLE_TH_ERRORS
-}
-
-typedef PyObject *(*getter)(PyObject *, void *);
-
-// NB: If you edit these properties/methods, update torch/_C/__init__.pyi.in
-
-static struct PyGetSetDef THPDevice_properties[] = {
-  {"type",       (getter)THPDevice_type, nullptr, nullptr, nullptr},
-  {"index",      (getter)THPDevice_index, nullptr, nullptr, nullptr},
-  {nullptr}
-};
-
-static PyMethodDef THPDevice_methods[] = {
-  {"__reduce__", THPDevice_reduce, METH_NOARGS, nullptr},
-  {nullptr}  /* Sentinel */
-};
-
-PyTypeObject THPDeviceType = {
-  PyVarObject_HEAD_INIT(nullptr, 0)
-  "torch.device",                        /* tp_name */
-  sizeof(THPDevice),                     /* tp_basicsize */
-  0,                                     /* tp_itemsize */
-  nullptr,                               /* tp_dealloc */
-  0,                                     /* tp_vectorcall_offset */
-  nullptr,                               /* tp_getattr */
-  nullptr,                               /* tp_setattr */
-  nullptr,                               /* tp_reserved */
-  (reprfunc)THPDevice_repr,              /* tp_repr */
-  nullptr,                               /* tp_as_number */
-  nullptr,                               /* tp_as_sequence */
-  nullptr,                               /* tp_as_mapping */
-  (hashfunc)THPDevice_hash,              /* tp_hash  */
-  nullptr,                               /* tp_call */
-  (reprfunc)THPDevice_str,               /* tp_str */
-  nullptr,                               /* tp_getattro */
-  nullptr,                               /* tp_setattro */
-  nullptr,                               /* tp_as_buffer */
-  Py_TPFLAGS_DEFAULT,                    /* tp_flags */
-  nullptr,                               /* tp_doc */
-  nullptr,                               /* tp_traverse */
-  nullptr,                               /* tp_clear */
-  (richcmpfunc)THPDevice_rc,             /* tp_richcompare */
-  0,                                     /* tp_weaklistoffset */
-  nullptr,                               /* tp_iter */
-  nullptr,                               /* tp_iternext */
-  THPDevice_methods,                     /* tp_methods */
-  nullptr,                               /* tp_members */
-  THPDevice_properties,                  /* tp_getset */
-  nullptr,                               /* tp_base */
-  nullptr,                               /* tp_dict */
-  nullptr,                               /* tp_descr_get */
-  nullptr,                               /* tp_descr_set */
-  0,                                     /* tp_dictoffset */
-  nullptr,                               /* tp_init */
-  nullptr,                               /* tp_alloc */
-  THPDevice_pynew,                       /* tp_new */
-};
-
-void THPDevice_init(PyObject *module)
-{
-  if (PyType_Ready(&THPDeviceType) < 0) {
-    throw python_error();
-  }
-  Py_INCREF(&THPDeviceType);
-  if (PyModule_AddObject(module, "device", (PyObject *)&THPDeviceType) != 0) {
-    throw python_error();
+    TORCH_CHECK_TYPE(
+      false, "device argument must be torch.device, str or int, not ",
+      py::type::of(device).attr("__name__").cast<std::string>()
+    );
   }
 }
+
+void initDeviceBindings(PyObject* module) {
+  py::options options;
+  options.disable_user_defined_docstrings();
+  options.disable_function_signatures();
+
+  // NB: If you edit these properties/methods, update torch/_C/__init__.pyi.in
+  py::class_<Device>(module, "device", py::is_final())
+      .def(py::init<const Device&>(), py::arg("device"))
+      .def(py::init(wrap_pybind_function(parseDevice)), py::arg("device"))
+      .def(py::init(
+          [](const std::string& type, DeviceIndex index) {
+            HANDLE_TH_ERRORS
+            auto check_device = Device(type);
+            TORCH_CHECK(
+              !check_device.has_index(),
+              "type (string) must not include an index because index was passed explicitly: ",
+              type
+            );
+            return parseDeviceIndex(check_device.type(), index);
+            END_HANDLE_TH_ERRORS_PYBIND
+          }),
+          py::arg("type"),
+          py::arg("index"))
+      .def(py::self == py::self)
+      .def(py::self != py::self)
+      .def(py::hash(py::self))
+      .def("__repr__",
+          [](const Device& device) {
+            std::ostringstream oss;
+            oss << "device(type=\'" << device.type() << "\'";
+            if (device.has_index()) {
+              // `device.index()` returns uint8_t which is treated as ascii while printing,
+              // hence casting it to uint16_t.
+              // https://stackoverflow.com/questions/19562103/uint8-t-cant-be-printed-with-cout
+              oss << ", index=" << static_cast<uint16_t>(device.index());
+            }
+            oss << ")";
+            return oss.str();
+          })
+      .def("__str__",
+          [](const Device& device) {
+            std::ostringstream oss;
+            oss << device;
+            return oss.str();
+          })
+      .def("__reduce__",
+          [](const Device& device) {
+            static auto device_cls = py::type::of<Device>();
+            auto type = deviceType(device);
+            auto args = device.has_index() ?
+              py::make_tuple(type, device.index()) :
+              py::make_tuple(type);
+            return py::make_tuple(device_cls, args);
+          })
+      .def_property_readonly("type", &deviceType)
+      .def_property_readonly(
+          "index",
+          [](const Device& device) -> c10::optional<DeviceIndex> {
+            if (device.has_index()) {
+              return device.index();
+            } else {
+              return c10::nullopt;
+            }
+          });
+}
+
+} // namespace torch