--- a/torch/csrc/Module.cpp
+++ b/torch/csrc/Module.cpp
@@ -730,7 +730,7 @@ PyObject* initModule() {
   THPLayout_init(module);
   THPMemoryFormat_init(module);
   THPQScheme_init(module);
-  THPDevice_init(module);
+  torch::initDeviceBindings(module);
   THPStream_init(module);
   ASSERT_TRUE(THPVariable_initModule(module));
   ASSERT_TRUE(THPFunction_initModule(module));