--- a/torch/csrc/jit/python/pybind_utils.cpp
+++ b/torch/csrc/jit/python/pybind_utils.cpp
@@ -69,10 +69,8 @@ IValue toIValue(py::handle obj, const TypePtr& type, c10::optional<int32_t> N) {
     }
     case TypeKind::StringType:
       return ConstantString::create(py::cast<std::string>(obj));
-    case TypeKind::DeviceObjType: {
-      auto device = reinterpret_cast<THPDevice*>(obj.ptr());
-      return device->device;
-    }
+    case TypeKind::DeviceObjType:
+      return py::cast<Device>(obj);
     case TypeKind::StreamObjType: {
       auto stream = reinterpret_cast<THPStream*>(obj.ptr());
       return static_cast<int64_t>(stream->cdata);