--- a/torch/csrc/utils/pybind.h
+++ b/torch/csrc/utils/pybind.h
@@ -5,6 +5,7 @@
 #include <ATen/ATen.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
+#include <pybind11/operators.h>
 
 #include <torch/csrc/DynamicTypes.h>
 #include <torch/csrc/autograd/python_variable.h>