--- a/torch/csrc/utils/python_arg_parser.h
+++ b/torch/csrc/utils/python_arg_parser.h
@@ -488,17 +488,8 @@ inline at::Device PythonArgs::device(int i) {
   if (!args[i]) {
     return at::Device(backendToDeviceType(dispatchKeyToBackend(torch::tensors::get_default_dispatch_key())));
   }
-  if (THPDevice_Check(args[i])) {
-    const auto device = reinterpret_cast<THPDevice*>(args[i]);
-    return device->device;
-  }
-  if (THPUtils_checkLong(args[i])) {
-    const auto device_index = THPUtils_unpackLong(args[i]);
-    TORCH_CHECK(device_index >= 0, "Device index must not be negative");
-    return at::Device(DeviceType::CUDA, device_index);
-  }
-  const std::string &device_str = THPUtils_unpackString(args[i]);
-  return at::Device(device_str);
+  // Original code does not mutate reference counts, so make a copy here
+  return torch::parseDevice(py::reinterpret_borrow<py::object>(args[i]));
 }
 
 inline at::Device PythonArgs::deviceWithDefault(int i, const at::Device& default_device) {