* `THPDevice_Check` and `THPDevice_new` stubs are still kept to ease the transition and reduce the number of files that needs to be changed.
* `torch::parseDevice` utility is introduced to parse Python `device` arguments. It accepts `torch.device`, `str` and `int`, just like `THPDevice_pynew`.