@lqf96 2021-01-07
Just pushed some small changes that move the bindings to `torch::python` namespace. I also make the `parseDevice` function accept a `py::handle` instead of a `py::object` for more efficiency.