@lqf96 2021-01-08
> you also might have to rebase to appease CI- I see an XLA test failing that looks like it's passing in master now

Yeah I can rebase this and kick off the test again. One thing to be aware of: we need to merge either this or #49771 first or there will be conflicts. After one of them is merged I'll rebase and update the other.