@robieta 2021-01-11
FYI: https://pytorch.org/tutorials/recipes/recipes/benchmark.html#collecting-instruction-counts-with-callgrind for collecting instruction counts.