@lqf96 2021-01-11
> Could you do some instruction count benchmarks using @robieta's Timer class (https://pytorch.org/docs/master/benchmark_utils.html) to verify that the new pybind11 isn't slower? (In particular, I'm wondering Device is properly being bound as a value type here)

Sure I can do that, but what should I benchmark? Creating devices? Getting devices from `torch.Generator`s and `torch.Tensor`s? Calling functions with `device` arguments? Or all of them?