@lqf96 2021-01-11
Ok I collected some timing data for a few examples that use `torch.device`:

Code for benchmarking:
<details>
<summary>Click to expand</summary>

```py
import torch
from torch.utils.benchmark import Timer

x = torch.ones(4, 5)
y = x.device

print(Timer(
    label="tensor-get-device",
    stmt="x.device",
    globals={"x": x}
).blocked_autorange())
print()

print(Timer(
    label="device-get-type",
    stmt="y.type",
    globals={"y": y}
).blocked_autorange())
print()

print(Timer(
    label="device-get-index",
    stmt="y.index",
    globals={"y": y}
).blocked_autorange())
print()

print(Timer(
    label="device-as-arg-1",
    stmt="x.to(y)",
    globals={"x": x, "y": y}
).blocked_autorange())
print()

print(Timer(
    label="device-as-arg-2",
    stmt="torch.arange(10, device=y)",
    globals={"torch": torch, "y": y}
).blocked_autorange())
print()
```
</details>

Master branch, debug build:
<details>
<summary>Click to expand</summary>

```
<torch.utils.benchmark.utils.common.Measurement object at 0x7f62577ebf40>
tensor-get-device
  Median: 147.28 ns
  IQR:    2.75 ns (145.95 to 148.69)
  136 measurements, 100000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f62eacccf10>
device-get-type
  Median: 438.05 ns
  IQR:    9.43 ns (432.67 to 442.11)
  458 measurements, 10000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f62577eb8e0>
device-get-index
  Median: 104.64 ns
  IQR:    1.90 ns (103.56 to 105.46)
  192 measurements, 100000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f62eaccc4c0>
device-as-arg-1
  Median: 3.41 us
  IQR:    0.08 us (3.37 to 3.45)
  588 measurements, 1000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f62577eb790>

device-as-arg-2
  Median: 24.14 us
  IQR:    0.68 us (23.73 to 24.41)
  832 measurements, 100 runs per measurement, 1 thread
```
</details>

This PR:
<details>
<summary>Click to expand</summary>

```
<torch.utils.benchmark.utils.common.Measurement object at 0x7f105ab37ac0>
tensor-get-device
  Median: 3.42 us
  IQR:    0.09 us (3.37 to 3.46)
  587 measurements, 1000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f105baff490>
device-get-type
  Median: 1.81 us
  IQR:    0.02 us (1.79 to 1.82)
  111 measurements, 10000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f105ab372b0>
device-get-index
  Median: 1.21 us
  IQR:    0.02 us (1.20 to 1.21)
  166 measurements, 10000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f105baff8b0>
device-as-arg-1
  Median: 4.41 us
  IQR:    0.10 us (4.36 to 4.46)
  454 measurements, 1000 runs per measurement, 1 thread

<torch.utils.benchmark.utils.common.Measurement object at 0x7f105ab37df0>
device-as-arg-2
  Median: 25.93 us
  IQR:    0.79 us (25.45 to 26.24)
  774 measurements, 100 runs per measurement, 1 thread
```
</details>

So it looks like there are some overhead caused by adopting pybind11, and the performance of calling `tensor.device` seems bad. It looks like copying `at::Device` might be the culprit, and I'll look into whether using `std::move` fixes the problem.